chore(searchable-dropdown): types in 1 file and event type for action event
